requirements.txt
flask
flask-cors
langchain
langchain-core
langchain-community
langchain-google-genai
python-dotenv
